wanna-build.yaml: add per distribution priority
wanna-build supports setting the distribution to 'any', meaning all the
supported distributions. This is useful for pybuildd which uses the api
version 2. However for that the priority computation needs to take into
account the distribution, as ordering is not done anymore at the buildd
level.

For now starts with this order of priority:
- *-security supported by the security team
- *-security supported by the LTS team
- (old-)stable
- testing
- sid (priority unchanged as it is the most used one)
- experimental
- *-backports*