Stop loading per buildd yaml file
This does not correspond anymore to the naming scheme we use for new
buildds, as they usually build for more than one architecture.

The per user yaml file should be used instead for the same result, which
has the advantage of being prefixed with "buildd-" with the naming
scheme we currently used.